Created the Editor class, merged the Cursor and Window classes into the Editor.
Added a function for cursor movement, it keeps you from leaving the
boundaries of the buffer.

Also added a "cursor_hint" that keeps track of your col position in a
line so you can move between lines and maintain the same general col
position when moving from a long line to a short line and then a long
line again. This is impossible to explain with words.

Ignore the "test.txt" file I'm using for testing the buffer, it's not
important and you shouldn't look at it.