Added a switchable cursor mode. Normal Mode for moving the cursor around and performing commands and shortcuts, and Insert Mode for inserting text.
Added the ability to insert text, finally! Also added with that the
ability to insert new lines and split existing lines with new line
characters.

Started laying the foundation for remappable keys with the Keys class
(keys.py) that holds all of the keys used for basically everything.

Lastly, created statusline.py for adding statuslines to buffers.
Statuslines are independant classes assigned to buffers, all it does
right now is returns a formatted string to the Editor which is then
printed at the very last row of the buffer, after all other lines are
drawn.