Manychanges, some breaking (for now)
Started colors + highlight group system, partially implemented. More
Highlight Groups will be implemented in the future, for now there is
just BufferText and LineNumbers, self explanatory. Colors are created in
curses using a custom rgb() function that converts the stupid 0-1000
values for rgb to 0-256, then color pairs for Highlight Groups are
initialized, and then applied to where they are supposed to be applied
to during refreshing. More to come on this later, I just started it
because I got sick of the line numbers being the same color as the rest
of the text.

Created popup.py, used for displaying popup windows on top of the Editor.
Windows are printed by the editor after the buffer. Windows are given a
row/col to be printed from, but their origin point can be set with
"anchor" to "top left", "bottom right", or "center". Windows will
automatically resize to fit content when created, and can be manually
resized at any point by calling their update function.

Slight modifications to buffer, editor, cursor, keys, and gem .py files
to interact with the above mentioned files.

Currently working on a bug with delete/backspace. All lines when read
from a file are appended with newline characters. A line with no text
still has a newline character, meaning it has a length of '1'
technically. If this newline character is deleted or backspaced, the
formatting breaks and everything becomes very unstable. Currently
working on testing out having the backspace/delete functions ignore the
newline characters or omitting the newline characters entirely, but
during initial testing this was found to have its own fair share of
bugs.