Many changes, more to come
Too many changes, probably. I should have done this in a smaller commit,
but I wanted to squash all of the bugs I knew of before committing
again. Then I got carried away with adding features on the side. So,
here we go.

Added more to `colors.py` and implemented all(?) of the existing
HighlightGroups to where they are supposed to be. The color system is an
absolute garbage fire and will need to be completely redone, once I can
think of a better way to do it. It's honestly disgusting right now.

Started work on `PopupInput` in `popup.py` which I'll likely use for the
prompt mode, once that's implemented. That'll be next, I think.
Currently unfinished.

Added a few new keybindings, just convenience things like jumping to the
beginning or end of a line, appending to the end of a line, etc. Nothing
major. I did, however, make it so backspace and delete work in normal
mode as well as insert mode.

Added saving and loading, arguably the most important feature of
anything that claims to be an editor. Kinda silly that I didn't have it
until now, it was simple to add, I just didn't feel like doing it until
now. You can also open a new file by passing in the name of a file that
doesn't exist or by not passing a name at all, however, there's no way
to name a file when saving at the moment, so it would probably not work
properly if you tried to save an unnamed file.

Pay no attention to NOTES.md, that's outdated. I keep forgetting to
update it before committing.

Pretty sure that's it.