Complete Rewrite Literally rewrote everything from scratch. I used some old code as a reference, but overall I cleaned up everything from the first draft, got some bugs worked out, make some things make more sense, added a few optimizations, and most importantly laid the groundwork for a much more modular system overall.
[Major changes]
Added `logger.py` which is just a simple logging system I wrote up
quick, all it does is prints to `gemlog.txt` when `log.write(message)` is
called. Useful for debugging, which is all I really made it for.

`Editor` was replaced with `Terminal`, it essentially serves the same
purpose, but is structured better and the name makes more sense.

Rather than the Terminal object doing all of the screen drawing operations, it
instead passes down the reference to the `stdsrc` to its' buffers.
This way, the Terminal class remains much smaller in size and tracing
bugs is easier because they report from different areas.

Buffers are more generic, they are made to be containers for Widgets,
which are also new. Right now, they're still a bit more particular and
specific than I want, and can only contain a text editor. That'll be
changing soon, and the text editor itself will be turned into a widget.
The statusbar will also be a widget. When a buffer is updated by it's
parent (the Terminal, of which there is only one), it's passed the
'stdscr' reference. Then the buffer updates all of its widgets. Rather
than a widget being passed the 'stdscr' so it can print itself, instead
a widget returns what it wants to be printed, and the buffer prints it.
This might change in the future, but it seems to work fine for now. I
made a stupidly simple "spinner" widget for testing out the system.

Another random note about buffers is that the cursor can be moved
between buffers, and acts inside of them independently, but this is
buggy and not fully implemented at the moment.

Popups have been remade from scratch completely. They aren't quite
finished yet, but they've been changed to act as sub-windows of the
Terminal, and they contain their own buffers, which contain their own
widgets, etc. This way, a popup window can be used for anything from a
simple block of text, up to a fully fledged text editor of its own, or
anything else that can be done with a standard buffer.

`keys.py` was replaced with `input.py` which is the same exact thing,
but has a better structure, and inputs are divided into subclasses,
like...

Leader Sequences! I'm really excited about this. You can define a
"Leader" key, and then create sequences like `[Key.Leader, 'q']`, which
is the sequence for exiting Gem. Now, whenever you're in Normal mode and
you press the Leader key, the input stream waits another 250ms for
another key to be entered. If you press Leader+q, the app quits! This is
especially convenient because you can define as many shortcuts as you
want without worrying about clashing with your OS or anything else.
Right now sequences are only two characters (Key.Leader plus another),
and the default Leader is space. Since sequences are only recognized in
Normal mode, it doesn't create a delay on the space key when inserting
text.

[Minor Changes]
Added a few more shortcuts, these are subject to change:
    - `(` Jump to the beginning of the line
    - ')' Jump to the end of the line
    - 'A' Jump to the end of the line and enter insert mode (A for "Append")
    - 'H' Jump to the end of the previous word on the line
    - 'L' Jump to the end of the next word on the line
    - 'J' Jump to the next blank line
    - 'K' Jump to the previous blank line
    - 'D' Delete the entire current line
    - 'Leader + s' Save the current file
    - 'Leader + q' Quit Gem

'statusline.py' was absorbed into 'buffer.py' but will eventually be
made into a widget.