Tons of fixes + changes More work done on `popup.py`, still not finished, but testing is going better. Many bugs fixed, the entire thing was basically rewritten a second time. You can now switch focus between the main buffer and a popup window with `w`, but not much else at the moment. More to come later on.
Fixed a few bugs in the history system. It's still not perfect, but it
gets the job done for now. Undo and redo are now seemingly fully
functional and free of bugs.

Minor changes to `buffer.py`, `terminal.py`, and `cursor.py` whilst
making changes to `popup.py`.

Implemented multi-key sequences! The biggest change of this commit. When
pressing a key, if that key is found in the list of single-key
shortcuts, it's processed and performed right away. However, if it's not
found in that list, Gem waits indefinitely for a second key to be
pressed, then sends both keys to be processed, allowing for some new
sequence shortcuts to be made. Added `gg` in place of `g` for
`BufferFirstRow` and `dd` in place of `D` for `LineDelete`.

Leader functionality is still the same with this new system, if the
Leader key is detected as the first key, it will wait 250ms for a second
key before moving on to processing. Since Leader key sequences are only
valid in Normal mode, and the spacebar (default Leader key) does nothing
in Normal mode, this is almost entirely unnoticable.