Add Prompt Mode and many small changes
Prompt Mode can be entered using the `ModePrompt` hotkey, default `:`
From there, commands can be typed and entered. Error handling has been
started but does not yet display errors on screen, instead it just
writes them to `gemlog.txt`. `cmd_handler.py` handles commands sent from
the Terminal and performs whatever actions it needs to from there. I
experimented with having the CmdHandler send the necessary calls back to
the Terminal, but this gets tricky when there are multiple functions
needing to be called with various arguments. I will likely revisit this
in the future. For now, the CmdHandler has access to the Terminal, and
from there can perform any actions that the Terminal can.

Popup Windows aren't yet perfect, but are good enough for now. I'll
likely end up reworking them more in the future, because they're still a
little bit wonky to work with, but for now I only wanted to get them
working enough to work with the prompt.

Fixed some bugs in the cursor, buffer, and terminal objects that were
uncovered during testing of the popup windows. Mostly minor things like
some numbers being slightly off or `cursor.goto()` not checking if the
destination is invalid.

Also added some more configuration to the buffers in order to work
better with popup windows. Things like border, title, line numbers,
empty line chars, and margins are now configurable. A buffer's ability
to be edited can also be toggled, which is necessary for popup windows
or buffers that are only meant to display information.

Lastly, commands. So far I've only added a few, and those are `save`,
`quit`, `savequit`, and `goto` as well as a quicker alternative to
`goto`. I'll detail these in the documentation, which I'm going to start
on right after this commit. All commands so far have various ways to be
called, mostly just shorthand or Vim compatibility. Since this editor is
basically a reimagining of Vim, I figured that's appropriate.

We're getting close to being "bootstrapped", by that I mean I could
start using Gem to develop Gem. That's sort of the "goal" I'm working
towards at the moment. This is all, of course, just for fun. This commit
alone was 7 days in the making, and had it's moments of frustration for
sure, but overall this has been a very fun project and I'm surprised it
all works as well as it does, regardless of the spaghetti-like nature of
some areas of the code.