Fix input stream bug
Single key inputs that had more than one possible input to check
for (like Escape or Backspace) weren't being checked for properly,
which caused the input stream to wait for a second key to be pressed
instead of moving forward. This would cause the next key you pressed to
be ignored, which got to be very annoying during testing, hence this
single tiny patch.