Partial fix for a bug with popup windows
Popup windows had to be a minimum of two lines if a border was enabled
so there was a line for the title and a line for the buffer contents.
Changed this so a single row popup now has increased height if it has
borders enabled. Also fixed a bug caused by the first change that was
incorrectly triggering a scroll down action in a single line popup.

Buffers now have `scrollable_v` and `scrollable_h` parameters to enable
or disable scrolling in both directions. I also added a `focusable`
parameter, but this is currently unused.