Rewrite of `buffer.update()`, plus addressing related bugs
I decided to rewrite `buffer.update()` because it was becoming a huge
mess. There were tons of bandaid fixes in there to make margins work
properly and I got sick of having to fix it every single time I changed
anything related to the popups or buffers, so I started it fresh.

Everything has been worked out now, and the margins are no longer an
issue. Everything works as expected. It turns out the biggest
contributer to the problems I was having was that the margins were being
calculated in the `init()` function that were based on attributes that
weren't being set until after `init()` had completed, a complete
oversight.