Implement toggleable LogViewer popup to Terminal
Added a new widget called `LogViewer` that is used in a special popup
window, created/destroyed with the commands `logviewer` or `lv` that
displays the most recent changes to `inklog.txt` which is generated (and
reset) every time ink is launched.

The log is something I've been using for the entire development period
so far, and it's been very helpful, but I wanted a way to be able to
view the log in real time.

Also fixed a small bug with updating cursor memory caused by a breaking
change I made in a previous commit, not sure when it happened. Fixed now
by implementing `cursor.update_memory()` which updates the cursor memory
of the cursor's current buffer whenever the cursor is moved or
`cursor.goto()` is called.