Add `delete()` and `erase()` methods to `logger.py`
`log.delete()` can be called to delete the log file. This is done after
a successful exit if enabled in Terminal, which it is by default.

`log.erase()` can be called to clear the log. This is done after
launching Ink to reset the log without having to delete it first.